refactor : remove '_common' folder of 'Article' folder - reason : '_common' folder is for shared component.  'deleteBtn' is not sharing component.
